requirements.txt
torch>=1.8.0
PyYAML>=5.0